+++ src/main/java/com/mopl/api/domain/review/repository/ReviewRepository.java
@@ -1,6 +1,7 @@
 package com.mopl.api.domain.review.repository;
 
 import com.mopl.api.domain.review.entity.Review;
+import com.mopl.api.domain.review.repository.impl.ReviewRepositoryCustom;
 import java.util.UUID;
 import org.springframework.data.jpa.repository.JpaRepository;
