+++ src/test/java/com/mopl/api/domain/review/service/ReviewServiceTest.java
@@ -11,7 +11,7 @@ import com.mopl.api.domain.content.exception.detail.ContentNotFoundException;
 import com.mopl.api.domain.review.exception.detail.ReviewAlreadyExistsException;
 import com.mopl.api.domain.review.exception.detail.ReviewNotFoundException;
 import com.mopl.api.domain.review.exception.detail.ReviewUnauthorizedException;
-import com.mopl.api.domain.user.exception.detail.UserNotFoundException;
+import com.mopl.api.domain.user.exception.user.detail.UserNotFoundException;
 import com.mopl.api.domain.review.mapper.ReviewMapper;
 import com.mopl.api.domain.review.repository.ReviewRepository;
 import com.mopl.api.domain.user.dto.response.UserDto;
@@ -29,7 +29,6 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.Arrays;
-import java.util.List;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
